CMake: fix build when tinycbor is used in another project via add_subdirectory
CMAKE_SOURCE_DIR always refers to the top level of the current CMake source tree (the top-level project source dir). When tinycbor is used in another project via add_subdirectory, CMAKE_SOURCE_DIR will refer to that another project's source dir, thus breaking the build.

To correctly support both cases (tinycbor is the top-level project vs. it is used in another top-level project), the CMAKE_SOURCE_DIR should be replaced with either CMAKE_CURRENT_SOURCE_DIR (which is "the source directory that is currently being processed") or alternatively with PROJECT_SOURCE_DIR (which is "the source directory of the last call to the project() command made in the current directory scope or one of its parents").

The same applies to the CMAKE_BINARY_DIR (should be replaced with CMAKE_CURRENT_BINARY_DIR or PROJECT_BINARY_DIR).

To be consistent with the current code, this commit replaces CMAKE_SOURCE_DIR in cmake/PackageConfig.cmake with PROJECT_SOURCE_DIR, as that file already uses PROJECT_ prefixed variables. In CMakeLists.txt CMAKE_SOURCE_DIR is replaced by CMAKE_SOURCE_DIR, and CMAKE_BINARY_DIR with CMAKE_CURRENT_BINARY_DIR.

Docs:

https://cmake.org/cmake/help/latest/variable/CMAKE_SOURCE_DIR.html
https://cmake.org/cmake/help/latest/variable/CMAKE_CURRENT_SOURCE_DIR.html
https://cmake.org/cmake/help/latest/variable/PROJECT_SOURCE_DIR.html

https://cmake.org/cmake/help/latest/variable/CMAKE_BINARY_DIR.html
https://cmake.org/cmake/help/latest/variable/CMAKE_CURRENT_BINARY_DIR.html
https://cmake.org/cmake/help/latest/variable/PROJECT_BINARY_DIR.html